PH-2: moved AuthActivity and MainActivity to folder ui
