+++ app/src/main/java/com/androiddev/profilehub/ui/main/MainActivity.kt
@@ -1,10 +1,11 @@
-package com.androiddev.profilehub
+package com.androiddev.profilehub.ui.main
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.androiddev.profilehub.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
