+++ app/src/main/res/layout/activity_auth.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".AuthActivity">
+    tools:context=".ui.auth.AuthActivity">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/authTitle"
